# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi System Laragpt",
        "description": "Descripción de mi API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/program_participants": {
            "get": {
                "tags": [
                    "Program Participants"
                ],
                "summary": "Show All Program Participants",
                "operationId": "1aad22350e7cbd46fd9d530de8cbe9a6",
                "responses": {
                    "200": {
                        "description": "List of ProgramParticipants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "program_id": {
                                                        "type": "integer",
                                                        "example": "My Program Id"
                                                    },
                                                    "entity_type": {
                                                        "type": "string",
                                                        "example": "My Entity Type any One of this 'App\\Models\\User', 'App\\Models\\Challenge', 'App\\Models\\Company'"
                                                    },
                                                    "entity_id": {
                                                        "type": "integer",
                                                        "example": "My Entity Id"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/program_participants?page=1"
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/program_participants?page=3"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/program_participants?page=2"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "24"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Program Participants"
                ],
                "summary": "Almacena un nuevo Program Participant en el sistema.",
                "operationId": "e1e8eedd2d07cb61ad76f295b0e3f835",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "program_id",
                                    "entity_type",
                                    "entity_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "program_id": {
                                        "type": "integer",
                                        "example": "My Program Id"
                                    },
                                    "entity_type": {
                                        "type": "string",
                                        "example": "My Entity Type any One of this 'App\\Models\\User', 'App\\Models\\Challenge', 'App\\Models\\Company'"
                                    },
                                    "entity_id": {
                                        "type": "integer",
                                        "example": "My Entity Id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Program Participant creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Entered Successfully"
                                        },
                                        "last_insert_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation exception"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/program_participants/{id}": {
            "get": {
                "tags": [
                    "Program Participants"
                ],
                "summary": "Show Specific Program Participant",
                "operationId": "e0041af2d785159c59ba4c4ddfffae21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "program_id": {
                                            "type": "integer",
                                            "example": "My Program Id"
                                        },
                                        "entity_type": {
                                            "type": "string",
                                            "example": "My Entity Type any One of this 'App\\Models\\User', 'App\\Models\\Challenge', 'App\\Models\\Company'"
                                        },
                                        "entity_id": {
                                            "type": "integer",
                                            "example": "My Entity Id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No exists Program Participant with id : #"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/program_participants/{program_participant}": {
            "put": {
                "tags": [
                    "Program Participants"
                ],
                "summary": "Update Exist Program Participant in Database",
                "operationId": "b03309f30da46ce59f3867e9950a7b54",
                "parameters": [
                    {
                        "name": "program",
                        "in": "path",
                        "description": "ID del program a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "program_id",
                                    "entity_type",
                                    "entity_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "program_id": {
                                        "type": "integer",
                                        "example": "My Program Id"
                                    },
                                    "entity_type": {
                                        "type": "string",
                                        "example": "My Entity Type any One of this 'App\\Models\\User', 'App\\Models\\Challenge', 'App\\Models\\Company'"
                                    },
                                    "entity_id": {
                                        "type": "integer",
                                        "example": "My Entity Id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Program Participant actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Update Successfully"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program Participant no encontrado"
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation exception"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Program Participants"
                ],
                "summary": "Remove Program Participant From System",
                "operationId": "5af78565aa03c4cab8f5f7acec7981ee",
                "parameters": [
                    {
                        "name": "program_participant",
                        "in": "path",
                        "description": "ID del program_participant a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Program Participant eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Delete Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Program no encontrado"
                    }
                }
            }
        },
...
    }
}